+++ docs/release_notes.rst
@@ -1,6 +1,13 @@
 Release Notes
 =============
 
+v1.1.4
+------
+
+* Added Chronos task support
+* Fixed ``privileged`` for Marathon output
+* Added support for Mesos fetcher
+
 v1.1.3
 ------